requirements.txt
# 数据处理
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
openpyxl>=3.1

# 机器学习与建模
scikit-learn>=1.3
xgboost>=2.0
shap>=0.43

# 可视化
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15
streamlit>=1.28

# 向量检索与嵌入
faiss-cpu>=1.7
sentence-transformers>=2.2

# LLM API
anthropic>=0.18
openai>=1.0

# 报告导出
python-docx>=0.8
markdown>=3.4

# 工具
tqdm>=4.65

# 测试
pytest>=7.4
